requirements.txt
stable-baselines3[extra]
gymnasium
gymnasium-minigrid
google-generativeai
python-dotenv
pandas
minigrid